camera bug fixed, hud refactoring, heavy anim begins.
